added a TopStats module for top 5 or 10 player stats for use in club and competition show pages
