+++ app/controllers/competitions_controller.rb
@@ -6,5 +6,14 @@ class CompetitionsController < ApplicationController
 
     def show
         @competition = Competition.find(params[:id])
+        @parent = @competition
+
+        @stats_array = [
+            "goals", "assists", "yellow_cards", "red_cards"
+        ]
+
+        @top_x = 10
+
+
     end
 end
