+++ app/models/competition.rb
@@ -1,3 +1,6 @@
 class Competition < ApplicationRecord
+    include TopStats::InstanceMethods
+
     has_many :clubs
+    has_many :players, through: :clubs
 end
