+++ app/models/player.rb
@@ -1,6 +1,7 @@
 class Player < ApplicationRecord
     belongs_to :position
     belongs_to :club
+    belongs_to :competition
     belongs_to :country
     has_one :playerStat
     has_many :goals
