+++ app/views/stats/_top_stats.html.erb
@@ -1,0 +1,6 @@
+<% @stats_array.each do |stat| %>
+  <h4>Most <%= stat.titleize %></h4>
+  <% @parent.getTopStat(stat, @top_x).each_with_index do |player, i|%>
+    <p><%= image_tag player[:image], size: "20x20" if @parent.class == Competition %>&nbsp;&nbsp;<%= i+1 %>. <%= player[:name]%>: <%= player[:num_stat] %></p>
+  <% end %>
+<% end %>
